refactor(v2): move provider ratings from header to metadata tab
- Remove ratings display from GameHeader component
- Move rating logic (formatRating, normalizeToTen, RatingChip type) to MetadataTab
- Add new 'Ratings' section in metadata tab (between Verification and Provider links)
- Display ratings with provider logos (16×16px) and formatted values
- Add 'ratings-label' translation key to all 18 locale files
- Improve organization: ratings now grouped with provider metadata instead of header
