+++ frontend/src/locales/bg_BG/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Публични бележки",
   "qr-scan-to-download": "Сканирай, за да свалиш",
   "rating": "Оценка",
+  "ratings-label": "Оценки",
   "redownload": "Свали отново",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Презапиши метаданните от всички избрани източници",
