+++ frontend/src/locales/cs_CZ/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Veřejné poznámky",
   "qr-scan-to-download": "Naskenujte pro stažení",
   "rating": "Hodnocení",
+  "ratings-label": "Hodnocení",
   "redownload": "Znovu stáhnout",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Přepsat metadata ze všech vybraných poskytovatelů",
