+++ frontend/src/locales/de_DE/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Öffentliche Notizen",
   "qr-scan-to-download": "Scannen zum Herunterladen",
   "rating": "Bewertung",
+  "ratings-label": "Bewertungen",
   "redownload": "Erneut herunterladen",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Metadaten aller ausgewählten Anbieter überschreiben",
