+++ frontend/src/locales/en_GB/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Public notes",
   "qr-scan-to-download": "Scan to download",
   "rating": "Rating",
+  "ratings-label": "Ratings",
   "redownload": "Re-download",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Overwrite metadata from all selected providers",
