+++ frontend/src/locales/es_ES/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Notas públicas",
   "qr-scan-to-download": "Escanear para descargar",
   "rating": "Valoración",
+  "ratings-label": "Valoraciones",
   "redownload": "Volver a descargar",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Sobrescribir metadatos de todos los proveedores seleccionados",
