+++ frontend/src/locales/fr_FR/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Notes publiques",
   "qr-scan-to-download": "Scanner pour télécharger",
   "rating": "Note",
+  "ratings-label": "Notes",
   "redownload": "Retélécharger",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Écraser les métadonnées de tous les fournisseurs sélectionnés",
