+++ frontend/src/locales/hu_HU/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Publikus jegyzetek",
   "qr-scan-to-download": "Olvasd be a letöltéshez",
   "rating": "Értékelés",
+  "ratings-label": "Értékelések",
   "redownload": "Újraletöltés",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Az összes kiválasztott szolgáltató metaadatainak felülírása",
