+++ frontend/src/locales/it_IT/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Note pubbliche",
   "qr-scan-to-download": "Scansiona per scaricare",
   "rating": "Valutazione",
+  "ratings-label": "Valutazioni",
   "redownload": "Riscarica",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Sovrascrivi i metadati da tutti i provider selezionati",
