+++ frontend/src/locales/ko_KR/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "공개 노트",
   "qr-scan-to-download": "스캔하여 다운로드",
   "rating": "평가",
+  "ratings-label": "평가",
   "redownload": "다시 다운로드",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "선택한 모든 제공자의 메타데이터로 덮어씁니다",
