+++ frontend/src/locales/pl_PL/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Notatki publiczne",
   "qr-scan-to-download": "Zeskanuj, aby pobrać",
   "rating": "Ocena",
+  "ratings-label": "Oceny",
   "redownload": "Pobierz ponownie",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Nadpisz metadane od wszystkich wybranych dostawców",
