+++ frontend/src/locales/pt_BR/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Notas públicas",
   "qr-scan-to-download": "Escaneie para baixar",
   "rating": "Avaliação",
+  "ratings-label": "Avaliações",
   "redownload": "Baixar novamente",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Sobrescrever metadados de todos os provedores selecionados",
