+++ frontend/src/locales/ro_RO/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Notițe publice",
   "qr-scan-to-download": "Scanează pentru a descărca",
   "rating": "Evaluare",
+  "ratings-label": "Evaluări",
   "redownload": "Redescarcă",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "Suprascrie metadatele de la toți furnizorii selectați",
