+++ frontend/src/locales/tr_TR/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "Herkese açık notlar",
   "qr-scan-to-download": "İndirmek için tarat",
   "rating": "Puan",
+  "ratings-label": "Puanlar",
   "redownload": "Yeniden indir",
   "redownload-manual": "Taranmış kılavuzu yeniden indir",
   "refresh-complete-desc": "Seçili tüm sağlayıcılardan meta veriyi üzerine yaz",
