+++ frontend/src/locales/zh_TW/rom.json
@@ -295,6 +295,7 @@
   "public-notes": "公共筆記",
   "qr-scan-to-download": "掃描以下載",
   "rating": "評分",
+  "ratings-label": "評分",
   "redownload": "重新下載",
   "redownload-manual": "Re-download scraped manual",
   "refresh-complete-desc": "覆寫所有所選提供商的元數據",
